feature(gameplay): refactor wing stats [#106]
