--- Assets/AddressableAssetsData/link.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b1d2736ad7e0ec34d8153d2470580aa6
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
