+++ Assets/Scenes/ScenesSystem/SystemManager.unity
@@ -3073,7 +3073,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &2064884057
 Transform:
   m_ObjectHideFlags: 0
