+++ Assets/Scripts/Manager/Character/CharacterComponentAIDuelCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2d5aa5b14ab20ce2aed8649347d6596
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
