+++ ProjectSettings/ProjectSettings.asset
@@ -146,8 +146,6 @@ PlayerSettings:
   - {fileID: -944628639613478452, guid: f3f3c069594d7cde29e79facb1593b7e, type: 3}
   - {fileID: 11400000, guid: ec11cdf0fd9de3ee0ac89dad99a1e8ea, type: 2}
   - {fileID: 11400000, guid: 4bd5f91fd934b8699928febed0a48ca3, type: 2}
-  - {fileID: -7716764870578761739, guid: 43af4fb01c7c80030966b19b89e7a2d1, type: 2}
-  - {fileID: 11400000, guid: b17b379a4ee545af1861693626dcd9dc, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
